+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/azure
 
+## 2.0.16
+
+### Patch Changes
+
+- Updated dependencies [68751f9]
+  - @ai-sdk/provider-utils@3.0.4
+  - @ai-sdk/openai@2.0.16
+
 ## 2.0.15
 
 ### Patch Changes