+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/gateway
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [68751f9]
+  - @ai-sdk/provider-utils@3.0.4
+
 ## 1.0.7
 
 ### Patch Changes